+++ tests/Unit/LogzIoFormatterTest.php
@@ -2,13 +2,14 @@
 
 namespace Tests\Unit;
 
-use DateTime;
 use Laravel\Logzio\Log\Formatter;
-use Monolog\Logger;
 use PHPUnit\Framework\TestCase;
+use Tests\Concerns\CanProvideLogRecord;
 
 class LogzioFormatterTest extends TestCase
 {
+    use CanProvideLogRecord;
+
     /**
      * Test it may be created with/without parameters
      *
@@ -30,35 +31,17 @@ class LogzioFormatterTest extends TestCase
     }
 
     /**
-     * Test it should format log record
+     * Test it should format a log record
      *
      * @return void
      */
     public function testItShouldFormatLog()
     {
         $formatter = new Formatter();
-        $record = $this->getRecord();
+        $record = $this->logRecord();
         $formatted = json_decode($formatter->format($record), true);
 
         $this->assertArrayHasKey('@timestamp', $formatted);
         $this->assertArrayNotHasKey('datetime', $formatted);
-    }
-
-    /**
-     * The simple log recore
-     *
-     * @return array
-     */
-    protected function getRecord($level = Logger::WARNING, $message = 'test', $context = []): array
-    {
-        return [
-            'message' => $message,
-            'context' => $context,
-            'level' => $level,
-            'level_name' => Logger::getLevelName($level),
-            'channel' => 'test',
-            'datetime' => DateTime::createFromFormat('U.u', sprintf('%.6F', microtime(true))),
-            'extra' => [],
-        ];
     }
 }